fix(rc): harden invite nullifiers and relay parsing
Reconciles the public-RC stack audit findings: consume invite enrollment nullifiers with register_if_new(), reject non-finite/deep/duplicate-key relay JSON responses before returning decoded payloads, remove circuit-breaker float aliases in favor of Decimal strings, validate BLS verifier env overrides, prefer packaged helpers before PATH, and normalize relay invite wall-clock state to integer Unix seconds.

Tests: .venv/bin/python -m pytest tests/test_gap_public_rc_invite_00b_enforcement_activation.py tests/test_invite_nullifier_registry_1576p.py tests/test_invite_nullifier_lmdb_persistence_1576r_fix1.py tests/test_phase_488_cdl_045_validator_circuit_breaker_surface_runtime.py tests/test_ilc_core_submodule_audit_fix1.py -q --tb=short; .venv/bin/python -m pytest tests/test_relay_client.py tests/test_relay_invite_shortcode.py tests/test_consensus_binary_deploy.py tests/test_relay_bls_rust_backend.py -q --tb=short

+++ ilc_core/network/relay/relay_server.py
@@ -1284,11 +1284,11 @@ class RelayRendezvousServer:
             raise RelayServerError("code_exhausted")
         return batch
 
-    def _unix_now(self) -> float:
+    def _unix_now(self) -> int:
         now = self._now_provider()
         if not isinstance(now, (int, float)) or isinstance(now, bool) or not math.isfinite(now):
             raise RelayServerError("relay_invite_clock_invalid")
-        return float(now)
+        return int(now)
 
     def handle_json_request(
         self,
